Remove abstract class ExtraSmallPeriodicFormatter
