--- src/Periodic/ExtraSmallPeriodicFormatter.php
@@ -1,46 +1,0 @@
-<?php
-
-namespace CultuurNet\CalendarSummaryV3\Periodic;
-
-use IntlDateFormatter;
-use CultuurNet\CalendarSummaryV3\Translator;
-
-abstract class ExtraSmallPeriodicFormatter
-{
-
-    /**
-     * @var IntlDateFormatter
-     */
-    protected $fmtDay;
-
-    /**
-     * @var IntlDateFormatter
-     */
-    protected $fmtMonth;
-
-    protected $trans;
-
-    public function __construct(string $langCode)
-    {
-        $this->fmtDay = new IntlDateFormatter(
-            $langCode,
-            IntlDateFormatter::FULL,
-            IntlDateFormatter::FULL,
-            date_default_timezone_get(),
-            IntlDateFormatter::GREGORIAN,
-            'd'
-        );
-
-        $this->fmtMonth = new IntlDateFormatter(
-            $langCode,
-            IntlDateFormatter::FULL,
-            IntlDateFormatter::FULL,
-            date_default_timezone_get(),
-            IntlDateFormatter::GREGORIAN,
-            'M'
-        );
-
-        $this->trans = new Translator();
-        $this->trans->setLanguage($langCode);
-    }
-}
